+++ graphql-maven-plugin-samples/graphql-maven-plugin-samples-Forum-server/src/main/java/org/forum/server/specific_code/PostPublisher.java
@@ -63,7 +63,7 @@ public class PostPublisher {
 	 * @param post
 	 */
 	void onNext(Post post) {
-		logger.debug("Emitting suscription notification for {}", post);
+		logger.trace("Emitting suscription notification for {}", post);
 		subject.onNext(post);
 	}