+++ graphql-maven-plugin-samples/graphql-maven-plugin-samples-allGraphQLCases-client/src/test/java/org/allGraphQLCases/subscription/SubscriptionCallbackListIntegerForTest.java
@@ -5,6 +5,9 @@ package org.allGraphQLCases.subscription;
 
 import java.util.List;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
 import com.graphql_java_generator.client.SubscriptionCallback;
 
 /**
@@ -14,27 +17,35 @@ import com.graphql_java_generator.client.SubscriptionCallback;
  */
 public class SubscriptionCallbackListIntegerForTest implements SubscriptionCallback<List<Integer>> {
 
+	/** The logger for this class */
+	static protected Logger logger = LoggerFactory.getLogger(SubscriptionCallbackListIntegerForTest.class);
+
+	final String clientName;
 	public List<Integer> lastReceivedMessage = null;
+
+	public SubscriptionCallbackListIntegerForTest(String clientName) {
+		this.clientName = clientName;
+	}
 
 	@Override
 	public void onConnect() {
-		// System.out.println("The subscription is connected");
+		logger.debug("The subscription is connected (for {})", clientName);
 	}
 
 	@Override
 	public void onMessage(List<Integer> t) {
-		// System.out.println("Received this list from the 'subscribeToAList' subscription: " + t);
+		logger.debug("Received this list from the 'subscribeToAList' subscription: {} (for {})", t, clientName);
 		lastReceivedMessage = t;
 	}
 
 	@Override
 	public void onClose(int statusCode, String reason) {
-		// System.out.println("The subscription is closed");
+		logger.debug("The subscription is closed (for {})", clientName);
 	}
 
 	@Override
 	public void onError(Throwable cause) {
-		System.out.println("Oups! An error occurred: " + cause.getMessage());
+		logger.error("Oups! An error occurred: " + cause.getMessage());
 	}
 
 }