+++ CHANGELOG.md
@@ -9,6 +9,15 @@ Whether the application uses the _graphql_, the _generateClientCode_ or the _gen
 * skipGenerationIfSchemaHasNotChanged: true _(both client and server mode)_
 
 
+# Not released yet
+
+Server mode:
+* Issue #70: The subscription notifications would not be properly sent when more than one client subscribed to a subscription.
+
+Client mode:
+* The response could not be deserialized, when it contains the (non standard) _extensions_ entry
+
+
 # 1.12.2
 
 Both modes (client and server):